Disable pregeneration for non-main namespaces
