+++ lib/mwUtil.js
@@ -2,8 +2,10 @@
 
 var uuid = require('cassandra-uuid').TimeUuid;
 var contentType = require('content-type');
-var HTTPError = require('hyperswitch').HTTPError;
 var jwt = require('jsonwebtoken');
+var HyperSwitch = require('hyperswitch');
+var HTTPError = HyperSwitch.HTTPError;
+var URI = HyperSwitch.URI;
 
 var mwUtil = {};
 
@@ -142,6 +144,28 @@ mwUtil.decodePagingToken = function(hyper, token) {
             }
         });
     }
+};
+
+var siteInfoCache = {};
+mwUtil.getSiteInfo = function(hyper, req) {
+    var rp = req.params;
+    if (!siteInfoCache[rp.domain]) {
+        siteInfoCache[rp.domain] = hyper.post({
+            uri: new URI([rp.domain, 'sys', 'action', 'siteinfo']),
+            body: {
+                siprop: 'general|namespaces|namespacealiases'
+            }
+        })
+        .then(function(res) {
+            return {
+                lang: res.body.query.general.lang,
+                legaltitlechars: res.body.query.general.legaltitlechars,
+                namespaces: res.body.query.namespaces,
+                namespacealiases: res.body.query.namespacealiases
+            };
+        });
+    }
+    return siteInfoCache[rp.domain];
 };
 
 module.exports = mwUtil;
