+++ sys/parsoid.js
@@ -6,6 +6,7 @@
 
 var P = require('bluebird');
 var HyperSwitch = require('hyperswitch');
+var Title = require('mediawiki-title').Title;
 var URI = HyperSwitch.URI;
 var HTTPError = HyperSwitch.HTTPError;
 
@@ -108,60 +109,68 @@ var PSP = ParsoidService.prototype;
 // TEMP TEMP TEMP!!!
 // Wiktionary / summary invalidation and mobileapps pregeneration
 PSP._dependenciesUpdate = function(hyper, req) {
-    var self = this;
     var rp = req.params;
-    var summaryPromise = P.resolve();
-    if (rp.domain.indexOf('wiktionary') === -1) {
-        // non-wiktionary, update summary
-        summaryPromise = hyper.get({
-            uri: new URI([rp.domain, 'v1', 'page', 'summary', rp.title]),
-            headers: {
-                'cache-control': 'no-cache'
+    return mwUtil.getSiteInfo(hyper, req)
+    .then(function(siteInfo) {
+        var rp = req.params;
+        var isMainNamespace = Title.newFromText(rp.title, siteInfo).getNamespace().isMain();
+        var updates = [];
+        if (isMainNamespace) {
+            var summaryPromise = P.resolve();
+            if (rp.domain.indexOf('wiktionary') === -1) {
+                // non-wiktionary, update summary
+                summaryPromise = hyper.get({
+                    uri: new URI([rp.domain, 'v1', 'page', 'summary', rp.title]),
+                    headers: {
+                        'cache-control': 'no-cache'
+                    }
+                });
+            } else if (/en.wiktionary/.test(rp.domain)) {
+                // wiktionary update, we are interested only in en.wiktionary
+                summaryPromise = hyper.get({
+                    uri: new URI([rp.domain, 'v1', 'page', 'definition', rp.title]),
+                    headers: {
+                        'cache-control': 'no-cache'
+                    }
+                });
             }
-        });
-    } else if (/en.wiktionary/.test(rp.domain)) {
-        // wiktionary update, we are interested only in en.wiktionary
-        summaryPromise = hyper.get({
-            uri: new URI([rp.domain, 'v1', 'page', 'definition', rp.title]),
-            headers: {
-                'cache-control': 'no-cache'
-            }
-        });
-    }
-    summaryPromise = summaryPromise.catch(function(e) {
-        if (e.status !== 501 && e.status !== 404) {
-            hyper.log('error/' + rp.domain.indexOf('wiktionary') < 0 ?
-            'summary' : 'definition', e);
+            summaryPromise = summaryPromise.catch(function(e) {
+                if (e.status !== 501 && e.status !== 404) {
+                    hyper.log('error/' + rp.domain.indexOf('wiktionary') < 0 ?
+                    'summary' : 'definition', e);
+                }
+            });
+            updates.push(summaryPromise);
         }
+
+        // Emit resource change events
+        var publicBaseURI = '//' + rp.domain + '/api/rest_v1/page';
+        updates.push(hyper.post({
+            uri: new URI([rp.domain, 'sys', 'events', '']),
+            body: [
+                { meta: { uri: publicBaseURI + '/html/' + encodeURIComponent(rp.title) } },
+                { meta: { uri: publicBaseURI + '/html/' + encodeURIComponent(rp.title)
+                + '/' + rp.revision } },
+                { meta: { uri: publicBaseURI + '/data-parsoid/' + encodeURIComponent(rp.title) } },
+                { meta: { uri: publicBaseURI + '/data-parsoid/' + encodeURIComponent(rp.title)
+                + '/' + rp.revision } },
+            ]
+        }));
+
+        if (isMainNamespace) {
+            updates.push(hyper.get({
+                uri: new URI([rp.domain, 'sys', 'mobileapps', 'mobile-sections', rp.title]),
+                headers: {
+                    'cache-control': 'no-cache',
+                }
+            }));
+        }
+
+        return P.all(updates);
+    })
+    .catch(function(e) {
+        hyper.log('warn/mobileapps', e);
     });
-
-
-    // Emit resource change events
-    var publicBaseURI = '//' + rp.domain + '/api/rest_v1/page';
-    var resourceChangePromise = hyper.post({
-        uri: new URI([rp.domain, 'sys', 'events', '']),
-        body: [
-            { meta: { uri: publicBaseURI + '/html/' + encodeURIComponent(rp.title) } },
-            { meta: { uri: publicBaseURI + '/html/' + encodeURIComponent(rp.title)
-                        + '/' + rp.revision } },
-            { meta: { uri: publicBaseURI + '/data-parsoid/' + encodeURIComponent(rp.title) } },
-            { meta: { uri: publicBaseURI + '/data-parsoid/' + encodeURIComponent(rp.title)
-                        + '/' + rp.revision } },
-        ]
-    });
-
-    return P.join(
-        resourceChangePromise,
-        summaryPromise,
-        hyper.get({
-            uri: new URI([rp.domain, 'sys', 'mobileapps', 'mobile-sections', rp.title]),
-            headers: {
-                'cache-control': 'no-cache',
-            }
-        }).catch(function(e) {
-            hyper.log('warn/mobileapps', e);
-        })
-    );
 };
 
 /**
